client/requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.3.0
Django==5.1.4
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-yasg==1.21.8
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
reportlab==4.2.5
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
